changes based on review comments
