+++ precompiles/erc20/query.go
@@ -11,14 +11,12 @@ import (
 	"strings"
 	"time"
 
-	errorsmod "cosmossdk.io/errors"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/x/authz"
 	authzkeeper "github.com/cosmos/cosmos-sdk/x/authz/keeper"
 	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
-	transfertypes "github.com/cosmos/ibc-go/v7/modules/apps/transfer/types"
+	"github.com/evmos/evmos/v16/ibc"
 	auth "github.com/evmos/evmos/v16/precompiles/authorization"
-	transferkeeper "github.com/evmos/evmos/v16/x/ibc/transfer/keeper"
 
 	"github.com/ethereum/go-ethereum/accounts/abi"
 	"github.com/ethereum/go-ethereum/common"
@@ -103,7 +101,7 @@ func (p Precompile) Decimals(
 ) ([]byte, error) {
 	metadata, found := p.bankKeeper.GetDenomMetaData(ctx, p.tokenPair.Denom)
 	if !found {
-		denomTrace, err := GetDenomTrace(p.transferKeeper, ctx, p.tokenPair.Denom)
+		denomTrace, err := ibc.GetDenomTrace(p.transferKeeper, ctx, p.tokenPair.Denom)
 		if err != nil {
 			return nil, ConvertErrToERC20Error(err)
 		}
@@ -213,30 +211,6 @@ func (p Precompile) Allowance(
 	return method.Outputs.Pack(allowance)
 }
 
-// GetDenomTrace returns the denomination trace from the corresponding IBC denomination. If the
-// denomination is not an IBC voucher or the trace is not found, it returns an error.
-func GetDenomTrace(
-	transferKeeper transferkeeper.Keeper,
-	ctx sdk.Context,
-	denom string,
-) (transfertypes.DenomTrace, error) {
-	if !strings.HasPrefix(denom, "ibc/") {
-		return transfertypes.DenomTrace{}, errorsmod.Wrapf(ErrNoIBCVoucherDenom, denom)
-	}
-
-	hash, err := transfertypes.ParseHexHash(denom[4:])
-	if err != nil {
-		return transfertypes.DenomTrace{}, err
-	}
-
-	denomTrace, found := transferKeeper.GetDenomTrace(ctx, hash)
-	if !found {
-		return transfertypes.DenomTrace{}, ErrDenomTraceNotFound
-	}
-
-	return denomTrace, nil
-}
-
 // GetAuthzExpirationAndAllowance returns the authorization, its expiration as well as the amount of denom
 // that the grantee is allowed to spend on behalf of the granter.
 func GetAuthzExpirationAndAllowance(
@@ -264,7 +238,7 @@ func GetAuthzExpirationAndAllowance(
 // getBaseDenomFromIBCVoucher returns the base denomination from the given IBC voucher denomination.
 func (p Precompile) getBaseDenomFromIBCVoucher(ctx sdk.Context, denom string) (string, error) {
 	// Infer the denomination name from the coin denomination base denom
-	denomTrace, err := GetDenomTrace(p.transferKeeper, ctx, denom)
+	denomTrace, err := ibc.GetDenomTrace(p.transferKeeper, ctx, denom)
 	if err != nil {
 		// FIXME: return 'not supported' (same error as when you call the method on an ERC20.sol)
 		return "", err
